Move incremental SMT theories into sub directory
So that the theory files are kept together and separated from the other
components of the incremental SMT decision procedure.
