+++ src/solvers/smt2_incremental/module_dependencies.txt
@@ -1,2 +1,3 @@
 solvers
+theories
 util
