+++ src/solvers/smt2_incremental/theories/smt_array_theory.cpp
