+++ src/solvers/smt2_incremental/theories/smt_bit_vector_theory.h
