+++ src/solvers/smt2_incremental/theories/smt_core_theory.h
