+++ unit/solvers/smt2_incremental/theories/module_dependencies.txt
@@ -1,3 +1,4 @@
 solvers/smt2_incremental/ast
 solvers/smt2_incremental/theories
 testing-utils
+util
