+++ unit/solvers/smt2_incremental/theories/smt_array_theory.cpp
@@ -2,11 +2,10 @@
 
 #include <util/mp_arith.h>
 
-#include <solvers/smt2_incremental/smt_array_theory.h>
-#include <solvers/smt2_incremental/smt_bit_vector_theory.h>
+#include <solvers/smt2_incremental/theories/smt_array_theory.h>
+#include <solvers/smt2_incremental/theories/smt_bit_vector_theory.h>
+#include <testing-utils/invariant.h>
 #include <testing-utils/use_catch.h>
-
-#include "testing-utils/invariant.h"
 
 TEST_CASE("SMT array theory \"select\".", "[core][smt2_incremental]")
 {
